refactor: 추가 꾸미기 반응형 리팩토링
