+++ src/components/template/SettingTemplate.tsx
@@ -48,11 +48,12 @@ export default SettingTemplate;
 const S = {
   Container: styled.section`
     position: relative;
-    min-width: 46.4rem;
     display: flex;
     flex-direction: column;
     @media ${device.tablet} {
-      margin-top: calc(${HEADER_HEIGHT.MOBILE} + 10.2rem);
+      margin-top: calc(${HEADER_HEIGHT.MOBILE} + 10rem);
+      padding-bottom: 10.2rem;
+      height: fit-content;
     }
     @media ${device.mobile} {
       display: flex;
@@ -86,7 +87,7 @@ const S = {
     @media ${device.tablet} {
       /* margin-bottom: 6.4rem; */
     }
-    @media ${device.mobile} {
+    @media ${device.tablet} {
       background: linear-gradient(to top, var(--background) 90%, transparent 100%);
       font-size: 1.6rem;
       position: fixed;
@@ -107,7 +108,7 @@ const S = {
     color: var(--gray600);
     width: 100%;
     text-align: center;
-    @media ${device.mobile} {
+    @media ${device.tablet} {
       width: calc(100% - 4rem);
       top: 1.7rem;
     }
