+++ src/pages/main/components/additionalSetup/SetIntro.tsx
